IntegerMap / LongMap: add put check contains in get...
AllocationImpl & CombinationService: modify toString

git-svn-id: https://repo.algotrader.ch/svn/trunk/AlgoTrader@4680 19eef36c-86f0-de11-afab-0004ac968479

+++ common/src/main/java/com/algoTrader/util/DoubleMap.java
@@ -23,9 +23,9 @@ public class DoubleMap<K> extends ConcurrentHashMap<K, Double> {
     public double getDouble(K key) {
 
         if (super.containsKey(key)) {
-            return super.get(key).longValue();
+            return super.get(key).doubleValue();
         } else {
-            return 0;
+            return 0.0;
         }
     }
 }
